[QLT] [1.0.10]
Continued the quality of the files affected by the refactoring of ActionManager

TODO:
- Finish the quality
- Finish the tests in squash

+++ railway-editor_java_version/Railway-editor/src/controller/KeyboardTool.java
@@ -25,7 +25,7 @@ public class KeyboardTool implements KeyListener {
 	@Override
 	public void keyPressed(KeyEvent e) {
 
-		if ((int) e.getKeyCode() == KeyEvent.VK_H) {
+		if (e.getKeyCode() == KeyEvent.VK_H) {
 			mainPanel.setHideHud(!mainPanel.isHideHud());
 			mainPanel.repaint();
 		}
